requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests
#^^ required to add dependencies for new packages

# if azure-functions doesn't install
# ^^ doesn't automatically resolve
# >> "

# if azure-functions doesn't install
# >> "python.exe -m pip install --upgrade pip" into root


#activate venv
#./.env/Scripts/activate